+++ barapost/barapost.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "3.11.f"
+__version__ = "3.11.g"
 # Year, month, day
-__last_update_date__ = "2020-02-28"
+__last_update_date__ = "2020-03-02"
 
 # |===== Check python interpreter version =====|