+++ barapost/prober.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
-__version__ = "1.17.e"
+__version__ = "1.17.f"
 # Year, month, day
-__last_update_date__ = "2020-02-28"
+__last_update_date__ = "2020-03-02"
 
 # |===== Check python interpreter version =====|